refactor: extract `getActionInfo` into a separate file
